+++ src/ts/ConvivaAnalytics.ts
@@ -1,4 +1,3 @@
-type ContentMetadata = Conviva.ContentMetadata;
 import {
   AdBreak,
   AdBreakEvent,
@@ -16,6 +15,7 @@ import { Html5Metadata } from './Html5Metadata';
 import { Html5Storage } from './Html5Storage';
 import { Html5Time } from './Html5Time';
 import { Html5Timer } from './Html5Timer';
+import { ContentMetadataBuilder, Metadata } from './ContentMetadataBuilder';
 
 type Player = PlayerAPI;
 
@@ -29,21 +29,6 @@ export interface ConvivaAnalyticsConfiguration {
    * production or automated testing.
    */
   gatewayUrl?: string;
-  /**
-   * A string value used to distinguish individual apps, players, locations, platforms, and/or deployments.
-   */
-  applicationName?: string;
-  /**
-   * A unique identifier to distinguish individual viewers/subscribers and their watching experience through
-   * Conviva's Viewers Module in Pulse.
-   * Can also be set in the source config of the player, which will take precedence over this value.
-   */
-  viewerId?: string;
-
-  /**
-   * A key-value map to send customer specific custom tags.
-   */
-  customTags?: { [key: string]: any };
 }
 
 export interface EventAttributes {
@@ -55,26 +40,21 @@ export interface ConvivaSourceConfig extends SourceConfig {
   contentId?: string;
 }
 
-export interface MetadataOverrides {
-  assetName?: string;
-}
-
 export class ConvivaAnalytics {
 
   private static readonly VERSION: string = '{{VERSION}}';
-
-  private player: Player;
+  private readonly player: Player;
   private events: typeof PlayerEvent;
-  private handlers: PlayerEventWrapper;
+  private readonly handlers: PlayerEventWrapper;
   private config: ConvivaAnalyticsConfiguration;
-  private contentMetadata: ContentMetadata;
+  private readonly contentMetadataBuilder: ContentMetadataBuilder;
   private sessionDataPopulated: boolean;
 
-  private systemFactory: Conviva.SystemFactory;
-  private client: Conviva.Client;
+  private readonly systemFactory: Conviva.SystemFactory;
+  private readonly client: Conviva.Client;
   private playerStateManager: Conviva.PlayerStateManager;
 
-  private logger: Conviva.LoggingInterface;
+  private readonly logger: Conviva.LoggingInterface;
   private sessionKey: number;
 
   /**
@@ -82,9 +62,6 @@ export class ConvivaAnalytics {
    * This flag is required because player.isAd() is unreliable and not always true between the events.
    */
   private isAd: boolean;
-
-  private playbackStarted: boolean;
-  private metadataOverrides: MetadataOverrides = {};
 
   constructor(player: Player, customerKey: string, config: ConvivaAnalyticsConfiguration = {}) {
     if (typeof Conviva === 'undefined') {
@@ -134,6 +111,7 @@ export class ConvivaAnalytics {
     }
 
     this.client = new Conviva.Client(clientSettings, this.systemFactory);
+    this.contentMetadataBuilder = new ContentMetadataBuilder(this.logger);
 
     this.registerPlayerEvents();
   }
@@ -146,19 +124,18 @@ export class ConvivaAnalytics {
    * relies on the players source we can't ensure that all metadata attributes are present at session creation.
    * Therefore it could be that there will be a 'ContentMetadata created late' issue after conviva validation.
    *
-   * @param assetName Will be used as contentMetadata.assetName if no source was loaded before. This overrides the
-   *                  source assetName. If no source was loaded and no assetName is present this method will throw an
-   *                  error.
+   * If no source was loaded and no assetName was set via updateContentMetadata this method will throw an error.
    */
-  public initializeSession(assetName?: string): void {
-    // This could be called before source loaded.
-    // Without setting the asset name on the content metadata when initializing the session the SDK will throw errors.
-    if (!this.player.getSource() && !assetName) {
-      throw('AssetName is missing. Provide assetName attribute or load player source first');
+  public initializeSession(): void {
+    if (this.isValidSession()) {
+      this.logger.consoleLog('There is already a session running.', Conviva.SystemSettings.LogLevel.WARNING);
+      return;
     }
 
-    if (assetName) {
-      this.metadataOverrides.assetName = assetName;
+    // This could be called before source loaded.
+    // Without setting the asset name on the content metadata the SDK will throw errors when we initialize the session.
+    if (!this.player.getSource() && !this.contentMetadataBuilder.assetName) {
+      throw('AssetName is missing. Load player source first or set assetName via updateContentMetadata');
     }
 
     this.internalInitializeSession();
@@ -204,6 +181,31 @@ export class ConvivaAnalytics {
     }
 
     this.client.sendCustomEvent(this.sessionKey, eventName, eventAttributes);
+  }
+
+  /**
+   * Will update the contentMetadata which are tracked with conviva.
+   *
+   * If there is an active session only permitted values will be updated and propagated immediately.
+   * If there is no active session the values will set on session creation.
+   *
+   * Attributes set via this method will override automatic tracked once.
+   * @param metadataOverrides Metadata attributes which will be used to track to conviva.
+   * @see ContentMetadataBuilder for more information about permitted attributes
+   */
+  public updateContentMetadata(metadataOverrides: Metadata) {
+    this.contentMetadataBuilder.setOverrides(metadataOverrides);
+
+    if (!this.isValidSession()) {
+      this.logger.consoleLog(
+        '[ Conviva Analytics ] no active session; Don\'t propagate content metadata to conviva.',
+        Conviva.SystemSettings.LogLevel.WARNING,
+      );
+      return;
+    }
+
+    this.buildContentMetadata();
+    this.updateSession();
   }
 
   /**
@@ -286,11 +288,10 @@ export class ConvivaAnalytics {
     this.playerStateManager.setPlayerType('Bitmovin Player');
     this.playerStateManager.setPlayerVersion(this.player.version);
 
-    this.contentMetadata = new Conviva.ContentMetadata();
     this.buildContentMetadata();
 
     // Create a Conviva monitoring session.
-    this.sessionKey = this.client.createSession(this.contentMetadata); // this will make the initial request
+    this.sessionKey = this.client.createSession(this.contentMetadataBuilder.build()); // this will make the initial request
 
     if (!this.isValidSession()) {
       // Something went wrong. With stable system interfaces, this should never happen.
@@ -307,58 +308,43 @@ export class ConvivaAnalytics {
    * Update contentMetadata which must be present before first video frame
    */
   private buildContentMetadata() {
-    this.contentMetadata.applicationName = this.config.applicationName || 'Unknown (no config.applicationName set)';
-    this.contentMetadata.viewerId = this.config.viewerId || null;
-    this.contentMetadata.duration = this.player.getDuration();
-    this.contentMetadata.streamType =
-      this.player.isLive() ? Conviva.ContentMetadata.StreamType.LIVE : Conviva.ContentMetadata.StreamType.VOD;
+    this.contentMetadataBuilder.duration = this.player.getDuration();
+    this.contentMetadataBuilder.streamType = this.player.isLive() ? Conviva.ContentMetadata.StreamType.LIVE : Conviva.ContentMetadata.StreamType.VOD;
 
-    let customMetadata: {} = {
+    this.contentMetadataBuilder.custom = {
       // Autoplay and preload are important options for the Video Startup Time so we track it as custom tags
       autoplay: PlayerConfigHelper.getAutoplayConfig(this.player) + '',
       preload: PlayerConfigHelper.getPreloadConfig(this.player) + '',
       integrationVersion: ConvivaAnalytics.VERSION,
-      ...this.config.customTags,
     };
 
     const source = this.player.getSource() as ConvivaSourceConfig;
+
     // This could be called before we got a source
     if (source) {
       this.buildSourceRelatedMetadata(source);
     }
-
-    // Either metadataOverrides or source should exist at this point
-    this.contentMetadata.assetName = this.metadataOverrides.assetName || source && this.getAssetNameFromSource(source);
-    this.contentMetadata.custom = customMetadata;
   }
 
   private buildSourceRelatedMetadata(source: ConvivaSourceConfig) {
-    this.contentMetadata.viewerId = source.viewerId || this.config.viewerId || null;
-    this.contentMetadata.custom = {
+    this.contentMetadataBuilder.assetName = this.getAssetNameFromSource(source);
+    this.contentMetadataBuilder.viewerId = source.viewerId || this.contentMetadataBuilder.viewerId;
+    this.contentMetadataBuilder.custom = {
       playerType: this.player.getPlayerType(),
       streamType: this.player.getStreamType(),
       vrContentType: source.vr && source.vr.contentType,
-      ...this.contentMetadata.custom,
+      ...this.contentMetadataBuilder.custom,
     };
 
-    // also include dynamic content metadata at initial creation
-    this.buildDynamicContentMetadata();
-  }
-
-  /**
-   * Update contentMetadata which are allowed during the session
-   */
-  private buildDynamicContentMetadata() {
-    const source = this.player.getSource();
-    this.contentMetadata.streamUrl = this.getUrlFromSource(source);
+    this.contentMetadataBuilder.streamUrl = this.getUrlFromSource(source);
   }
 
   private updateSession() {
     if (!this.isValidSession()) {
       return;
     }
-    this.buildDynamicContentMetadata();
-    this.client.updateContentMetadata(this.sessionKey, this.contentMetadata);
+
+    this.client.updateContentMetadata(this.sessionKey, this.contentMetadataBuilder.build());
   }
 
   private getAssetNameFromSource(source: ConvivaSourceConfig): string {
@@ -387,6 +373,7 @@ export class ConvivaAnalytics {
 
     this.sessionKey = Conviva.Client.NO_SESSION_KEY;
     this.sessionDataPopulated = false;
+    this.contentMetadataBuilder.reset();
   };
 
   private isValidSession(): boolean {
@@ -425,7 +412,7 @@ export class ConvivaAnalytics {
     }
 
     this.buildSourceRelatedMetadata(this.player.getSource());
-    this.client.updateContentMetadata(this.sessionKey, this.contentMetadata);
+    this.updateSession();
   };
 
   private onPlay = (event: PlaybackEvent) => {
@@ -442,7 +429,7 @@ export class ConvivaAnalytics {
   };
 
   private onPlaying = (event: PlaybackEvent) => {
-    this.playbackStarted = true;
+    this.contentMetadataBuilder.setPlaybackStarted(true);
     this.debugLog('playing', event);
     this.updateSession();
     this.onPlaybackStateChanged(event);
@@ -450,8 +437,7 @@ export class ConvivaAnalytics {
 
   // When the first ON_TIME_CHANGED event arrives, the loading phase is finished and actual playback has started
   private onTimeChanged = (event: PlaybackEvent) => {
-    if (this.isValidSession() && !this.playbackStarted) {
-      // fallback for player versions <= 7.2 which do not support ON_PLAYING Event
+    if (this.isValidSession()) {
       this.onPlaying(event);
     }
   };
@@ -617,20 +603,20 @@ class PlayerConfigHelper {
 
     if (BrowserUtils.isMobile()) {
       if (playerConfig.adaptation
-          && playerConfig.adaptation.mobile
-          && playerConfig.adaptation.mobile.preload !== undefined) {
+        && playerConfig.adaptation.mobile
+        && playerConfig.adaptation.mobile.preload !== undefined) {
         return playerConfig.adaptation.mobile.preload;
       }
     } else {
       if (playerConfig.adaptation
-          && playerConfig.adaptation.desktop
-          && playerConfig.adaptation.desktop.preload !== undefined) {
+        && playerConfig.adaptation.desktop
+        && playerConfig.adaptation.desktop.preload !== undefined) {
         return playerConfig.adaptation.desktop.preload;
       }
     }
 
     if (playerConfig.adaptation
-        && playerConfig.adaptation.preload !== undefined) {
+      && playerConfig.adaptation.preload !== undefined) {
       return playerConfig.adaptation.preload;
     }